Implement `dialogue::enter`
